connect firestore data to event recyclerview
