+++ app/src/main/java/com/snc/farmaccount/home/DayViewModel.kt
@@ -1,8 +1,13 @@
 package com.snc.farmaccount.home
 
+import android.util.Log
 import androidx.lifecycle.*
+import com.google.firebase.firestore.FirebaseFirestore
+import com.google.firebase.firestore.FirebaseFirestoreSettings
 import com.snc.farmaccount.`object`.Event
+import java.text.SimpleDateFormat
 import java.util.*
+import kotlin.collections.ArrayList
 
 class DayViewModel: ViewModel() {
 
@@ -12,8 +17,9 @@ class DayViewModel: ViewModel() {
     private var week:Int = 0
     private var weekName = ""
     var date = MutableLiveData<String>()
-    var pickDate = MutableLiveData<String>()
-    var eventList = MutableLiveData<List<Event>>()
+    private var pickDate = MutableLiveData<String>()
+    lateinit var firebaseEvent : Event
+    var dataList = ArrayList<Event>()
 
     private val _event = MutableLiveData<List<Event>>()
     val event: LiveData<List<Event>>
@@ -26,8 +32,41 @@ class DayViewModel: ViewModel() {
         week()
     }
 
-    fun getEvent() {
-        _event.value = eventList.value
+    fun getFirebase() {
+        val db = FirebaseFirestore.getInstance()
+        val settings = FirebaseFirestoreSettings.Builder()
+            .setTimestampsInSnapshotsEnabled(true)
+            .build()
+        db.firestoreSettings = settings
+        db.collection("User")
+            .get()
+            .addOnCompleteListener { task ->
+                if (task.isSuccessful) {
+                    for (document in task.result!!) {
+                        val sub = FirebaseFirestore.getInstance()
+                        sub.collection("User/${document.id}/Event")
+                            .get()
+                            .addOnCompleteListener { task ->
+                                if (task.isSuccessful) {
+                                    for (document in task.result!!) {
+                                        val timestamp = document["date"] as com.google.firebase.Timestamp
+                                        val milliseconds = timestamp.seconds * 1000 + timestamp.nanoseconds / 1000000
+                                        val sdf = SimpleDateFormat("yyyy.MM.dd (EEEE)")
+                                        val netDate = Date(milliseconds)
+                                        val date = sdf.format(netDate).toString()
+                                        firebaseEvent = document.toObject(Event::class.java)
+                                        dataList.add(firebaseEvent)
+                                    }
+
+                                } else {
+                                    Log.w("Sophie_db_fail", "Error getting documents.", task.exception)
+                                }
+                                _event.value = dataList
+                                Log.w("Sophie_db_list", "$dataList")
+                            }
+                    }
+                }
+            }
     }
 
     private fun week() {
