+++ app/src/main/java/com/snc/farmaccount/object/Event.kt
@@ -3,10 +3,10 @@ package com.snc.farmaccount.`object`
 import java.util.*
 
 data class Event (
-    var id: Long,
-    var price: String,
-    var tag: String,
-    var description: String,
+    var id: Long ?= null,
+    var price: String ?= null,
+    var tag: String ?= null,
+    var description: String ?= null,
     var date: Date = Calendar.getInstance().time,
-    var status: Boolean
+    var status: Boolean ?= null
 )
